+++ Sources/Presentation/UI/SignedDocument/SignedDocumentView.swift
@@ -62,7 +62,7 @@ struct SignedDocumentView<Router: RouterGraph>: View {
     .toolbar {
       ToolbarItem(placement: .navigationBarTrailing) {
         Button("Done") {
-          viewModel.pause()
+          viewModel.onPause()
         }
       }
     }
